feat: 터미널 컴포넌트 삽입
