+++ web/react-app/src/App.js
@@ -18,8 +18,7 @@ function App() {
         <nav>
           <Link to="/">Home</Link> | <Link to="/about">About</Link> |{' '}
           <Link to="/Counter">Counter</Link> | <Link to="/Input">Input</Link> |{' '}
-          <Link to="/List">List</Link> | <Link to="/ExecuteCommand">H-execute</Link>|{' '}
-          <Link to="/Train">Train</Link>
+          <Link to="/List">List</Link> | <Link to="/Train">Train</Link>
         </nav>
         <Switch>
           <Route path="/" exact component={Home} />
